refactor: Pull out function to get account number
