+++ tests/conftest.py
@@ -10,7 +10,6 @@ from mypy_boto3_s3 import S3Client
 from mypy_boto3_s3control import S3ControlClient
 from mypy_boto3_ssm import SSMClient
 from mypy_boto3_stepfunctions import SFNClient
-from mypy_boto3_sts import STSClient
 
 basicConfig(level=INFO)
 logger = getLogger(__name__)
@@ -34,11 +33,6 @@ def s3_control_client() -> S3ControlClient:
 @pytest.fixture()
 def ssm_client() -> SSMClient:
     return boto3.client("ssm")
-
-
-@pytest.fixture()
-def sts_client() -> STSClient:
-    return boto3.client("sts")
 
 
 @pytest.fixture()
