+++ tests/test_step_function_logging.py
@@ -24,9 +24,9 @@ class TestLogging:
         expected_response_log = dumps({S3_BATCH_RESPONSE_KEY: s3_batch_response})
 
         with patch.object(self.logger, "debug") as logger_mock, patch(
-            "backend.step_function.STS_CLIENT.get_caller_identity"
-        ) as sts_mock:
-            sts_mock.return_value = {"Account": any_account_id()}
+            "backend.step_function.get_account_number"
+        ) as get_account_number_mock:
+            get_account_number_mock.return_value = any_account_id()
 
             # When
             get_s3_batch_copy_status("test")
